requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.4.0
beautifulsoup4==4.10.0
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.12
discord.py==1.7.3
idna==3.3
multidict==6.0.2
numpy==1.22.3
opencc-python-reimplemented==0.1.6
pandas==1.4.1
PyMySQL==1.0.2
pypinyin==0.46.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2022.1
requests==2.27.1
six==1.16.0
soupsieve==2.3.1
typing_extensions==4.1.1
urllib3==1.26.9
Wikipedia-API==0.5.4
yarl==1.7.2